shrink some of InsuranceClaim fields

+++ programs/clearing_house/src/instructions/admin.rs
@@ -1226,9 +1226,9 @@ pub fn handle_update_perp_market_margin_ratio(
 )]
 pub fn handle_update_perp_market_max_imbalances(
     ctx: Context<AdminUpdatePerpMarket>,
-    unrealized_max_imbalance: u128,
-    max_revenue_withdraw_per_period: u128,
-    quote_max_insurance: u128,
+    unrealized_max_imbalance: u64,
+    max_revenue_withdraw_per_period: u64,
+    quote_max_insurance: u64,
 ) -> Result<()> {
     let perp_market = &mut load_mut!(ctx.accounts.perp_market)?;
